+++ src/tradelayer/tradelayer.cpp
@@ -1044,7 +1044,7 @@ int ParseTransaction(const CTransaction& tx, int nBlock, unsigned int idx, CMPTr
  {
      bool count = false;
      uint64_t nvalue = 0;
-     
+
      if (property != LTC) return count;
 
      for (unsigned int n = 0; n < tx.vout.size(); ++n)
@@ -5278,7 +5278,7 @@ void mastercore::update_sum_upnls()
 }
 
 /* margin needed for a given position */
-int64_t mastercore::pos_margin(uint32_t contractId, const std::string& address, uint32_t margin_requirement)
+int64_t mastercore::pos_margin(uint32_t contractId, const std::string& address, uint64_t margin_requirement)
 {
         arith_uint256 maintMargin;
